Grown man business here. Just needed to make a few modifications to the `exploration_view.py` file for clarity and better cohesiveness in what was being written to ensure that the model could truly understand what we're trying to convey view 'few-shot learning' (which is clearly what's being implemented here).
Also, I swapped out the gpt-3.5-turbo model for gpt-4. That does mean that we're restricted in our context window that we can use here (down to 8k tokens if you're one of the members of the proletariat like I am). But never fear because  that should be more than enough since this entire program is designed for a single-turn conversation.

In essence, the user is to:

1. Give the prompt to the model
2. The model then distills that prompt into a generalized task description.
3. That task description is what is then used to decompose the task itself into a series of subtasks that each serve as the constituent facets of the overall prompt.
4. Each sub-task is really more like a sub-prompt that's meant to be fed to an LLM.
5. The goal here is that by feeding each subtask to the LLM and then having the LLM generate a response, we can aggregate those responses and then use them to generate a response to the original prompt.

This also sparks my idea as it pertains to 'Graph of Thought' analysis. This is really the best way for anybody to implement 'Graph of Thought' into their own pipeline because that essentially involves one breaking down a prompt into constituent tasks and then having those tasks be handled separately (or in parallel to one another). As this is taking place, those prompts are transformed into responses that exact in various states of being (initial, intermediate, final, etc.).  It is up to the algorithm how it ultimately amalgamates those responses into one cogent answer - but this is something that's already accounted for in the actual codebase of that project.

So when I have time (and after I'm finished here), I'm going to merge the modified code that I have here into the 'Graph of Thoughts' repo (via fork) so that we can have one cohesive pipeline for this entire mini-project (backend prompt engineering since the user should be none the wiser as to what's gong on behind the scenes unless they truly feel curious enough to look into it).